Fix ReadAt with offset==EOF
Fix a problem when a `ReadAt` requests a read at the offset that in plaintext corresponds exactly to a block starting at the EOF.

We forward to where the block *would* be, thereby we never read the previous block header, and therefore do not get the "final" marker set

We fix by reading (and skipping) the previous block if a request starts exactly at a block boundary to pick up the "final" marker.

We calculate `t`, to be the number of full blocks to skip and `k` to be the number of bytes to skip before the response starts:

```
	t := offset / int64(maxPayloadSize)
	k := offset % int64(maxPayloadSize)
```

We add the following if we start at a block boundary after the first block:

```
	if offset > 0 && k == 0 {
		k = maxPayloadSize
		t--
	}
```

This will read the previous block to read the "final" flag and skip its content.